Removed redundancy in editable shape implementations
